+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project should be documented in this file.
 - A `FunctionPatcher` mutator to attack function versioning and inlining caches, by @devdanzin.
 - A `TraceBreaker` mutator to attack the JIT's ability to form superblocks using trace-unfriendly code, by @devdanzin.
 - An `ExitStresser` mutator to attack the JIT's side-exit mechanism with loops containing many frequently taken branches, by @devdanzin.
+- A `DeepCallMutator` to attack the JIT's trace stack limit by adding a chain of nested function calls, by @devdanzin.
 
 
 ### Enhanced